Delete LICENSE-ASL-2.0
Deleting the Apache v2.0 license from our codebase. It appears to have been added erroneously on project creation, perhaps in emulation of https://github.com/DataBiosphere/leonardo. If one of our dependencies is licensed under Apache v2.0, we should add this back to cover redistribution, and note in the repo which piece of code this covers. But unless and until that happens, this commit removes the extraneous license.