use cached download prices (#310)
This replaces the (very limited) API key we were using with a cached version of the data that we'll be updating nightly.

This is being done because we want to avoid the appearance of having exposed a secret. While putting the key in the repo was intentional, because we estimated that the possibility and impact of any potential misuse was exceedingly small, it was observed that this could appear sloppy, and we want to avoid that.