[CA-1587] Paint Icon for Status of Billing Account Change (#2594)
RR: https://broadworkbench.atlassian.net/browse/CA-1487

When a user changes the Google Billing Account associated with a Terra Billing Project, Rawls updates the Billing Accounts of the Billing Project's Workspaces' Google Projects asynchronously.
While a user waits for these changes to be made, we will paint an icon next to each workspace that best matches the status of the change. See the mock-up in the ticket for more information.

In this change:

Add padding to the Billing Project's Workspace table headings for an icon beside each workspace
Conditionally add an icon beside each workspace based upon the status of the update. If all workspaces have been updated successfully, don't draw any icon.