[WOR-768] Extract workspace modal state to hook (#3783)
This is an attempt at extracting a React hook to contain the business logic of the DeleteWorkspaceModal. We are going to be embarking on some more substantial changes to remediate Azure workspace deletion issues, so the idea is to get some amount of test coverage on this code before going any further.

The functionality of the modal should remain unchanged after this PR, this is a straight refactor. I've tested locally against both Azure and GCP workspaces and things are working ok (including the stateful bits like the bytes in bucket and collaborator email callouts in the modal).