Do not refresh submissions when viewing all
When a user viewed all submissions, the refresh
was scheduled with the original "30 days" filter
so if there were any running submissions the page
would refresh to only show the last 30 days
(while "all" was still selected/displayed).

Now, the UI does not refresh running submissions if
all submissions are selected.